+++ python/ambassador/ir/irratelimit.py
@@ -82,6 +82,7 @@ class IRRateLimit(IRFilter):
             "timeout_ms": config.get("timeout_ms", 20),
             "request_type": "both",  # XXX configurability!
             "failure_mode_deny": config.get("failure_mode_deny", False),
+            "rate_limited_as_resource_exhausted": config.get("rate_limited_as_resource_exhausted", False),
         }
 
         self.sourced_by(config)