+++ pkg/api/envoy/config/core/v3/protocol.pb.validate.go
@@ -1191,8 +1191,6 @@ func (m *Http1ProtocolOptions) validate(all bool) error {
 		}
 	}
 
-	// no validation rules for AllowCustomMethods
-
 	if len(errors) > 0 {
 		return Http1ProtocolOptionsMultiError(errors)
 	}