+++ CHANGELOG.md
@@ -90,7 +90,7 @@ it will be removed; but as it won't be user-visible this isn't considered a brea
 
 ### Emissary-ingress and Ambassador Edge Stack
 
-## [3.8.0] TBD
+## [3.8.0] August 29, 2023
 [3.8.0]: https://github.com/emissary-ingress/emissary/compare/v3.7.2...v3.8.0
 
 ### Emissary-ingress and Ambassador Edge Stack
@@ -145,6 +145,12 @@ it will be removed; but as it won't be user-visible this isn't considered a brea
 - Feature: This upgrades Emissary-ingress to be built on Envoy v1.26.1. This provides
   Emissary-ingress with the latest security patches, performances enhancments, and features offered
   by the envoy proxy.
+
+- Feature: By default, Envoy will return an `UNAVAILABLE` gRPC code when a request is rate limited.
+  The `RateLimitService` resource now exposes the <a
+  href="https://www.envoyproxy.io/docs/envoy/v1.26.0/configuration/http/http_filters/rate_limit_filter">use_resource_exhausted_code</a>
+  option. Set `grpc.use_resource_exhausted_code: true` so Envoy will return a `RESOURCE_EXHAUSTED`
+  gRPC code instead.
 
 ## [3.6.0] April 17, 2023
 [3.6.0]: https://github.com/emissary-ingress/emissary/compare/v3.5.0...v3.6.0